requirements.txt
anthropic>=0.92.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
